fix : 렌더러의 뎁스 스테이트를 잘못 생성하고 있던 문제 수정, 헤더파일관련 에러 수정
