+++ Engine/Source/CoreUObject/Components/SceneComponent.h
@@ -1,12 +1,13 @@
 ﻿#pragma once
 
-#include "Core/Math/Vector.h"
-#include "Core/Math/Matrix.h"
-#include "Core/Math/Transform.h"
 #include "Core/Container/Set.h"
-#include "CoreUObject/Object.h"
 #include "CoreUObject/Components/ActorComponent.h"
-#include "Core/Math/Box.h"
+#include "Core/Math/Transform.h"
+
+struct FVector;
+struct FMatrix;
+struct FTransform;
+struct FBox;
 
 class USceneComponent : public UActorComponent
 {
